Create brick-problem.py
These are the type of logic problems we will be doing. You won't have google, so don't google the answer. You can work together, but it will not do you any good to google. This took me a bit to solve by myself, so it will take some trial and error. 


Making bricks. This is a challenge. Not a grade.

We want to make a row of bricks that is goal inches long. 

We have a number of small bricks (1 inch each) and big bricks (5 inches each). 

Return True if it is possible to make the goal by choosing from the given bricks. 

This is a little harder than it looks and can be done without any loops. 

Examples:

make_bricks(3, 1, 8) → true
make_bricks(3, 1, 9) → false
make_bricks(3, 2, 10) → true